Split Net modules into per-command modules
