+++ test/toolshed/ns_lookup_test.exs
@@ -1,0 +1,8 @@
+defmodule Toolshed.NslookupTest do
+  use ExUnit.Case
+  import Toolshed.Nslookup
+
+  test "nslookup/1 returns correct value" do
+    assert nslookup("google.com") == :"do not show this result in output"
+  end
+end
